factored out redundant pose sampling code
